+++ packages/ckeditor5-html-support/src/generalhtmlsupport.js
@@ -9,9 +9,9 @@
 
 import { Plugin } from 'ckeditor5/src/core';
 import DataFilter from './datafilter';
-import CodeBlockHtmlSupport from './integrations/codeblock';
 import TableElementSupport from './integrations/table';
-import DualContentModelElementSupport from './integrations/dualcontentmodelelementsupport';
+import CodeBlockElementSupport from './integrations/codeblock';
+import DualContentModelElementSupport from './integrations/dualcontent';
 
 /**
  * The General HTML Support feature.
@@ -44,8 +44,8 @@ export default class GeneralHtmlSupport extends Plugin {
 	static get requires() {
 		return [
 			DataFilter,
-			CodeBlockHtmlSupport,
 			TableElementSupport,
+			CodeBlockElementSupport,
 			DualContentModelElementSupport
 		];
 	}
