+++ packages/ckeditor5-html-support/src/integrations/codeblock.js
@@ -18,7 +18,7 @@ import DataFilter from '../datafilter';
  *
  * @extends module:core/plugin~Plugin
  */
-export default class CodeBlockHtmlSupport extends Plugin {
+export default class CodeBlockElementSupport extends Plugin {
 	static get requires() {
 		return [ DataFilter ];
 	}
