+++ packages/ckeditor5-html-support/src/integrations/dualcontent.js
